refactor: rename settings dialog
