+++ assets/ui/settings_dialog.ui
