+++ docs/development.md
@@ -20,7 +20,7 @@ All the required tools are available by installing `pyside6`:
 * `.ui > .py` Converter: `pyside6-uic`
   At the project root (use full paths if relative doesn't work):
   ```
-  pyside6-uic assets/ui/settings.ui -o hd_active/ui/forms/settings_ui.py
+  pyside6-uic assets/ui/settings.ui -o hd_active/ui/forms/ui_settings_dialog.py
   ```
 
 !!! Note
