+++ src/hd_active/ui/forms/ui_settings_dialog.py
@@ -1,60 +1,23 @@
 # -*- coding: utf-8 -*-
 
 ################################################################################
-## Form generated from reading UI file 'settings.ui'
+## Form generated from reading UI file 'settings_dialog.ui'
 ##
-## Created by: Qt User Interface Compiler version 6.2.4
+## Created by: Qt User Interface Compiler version 6.7.2
 ##
 ## WARNING! All changes made in this file will be lost when recompiling UI file!
 ################################################################################
 
-from PySide6.QtCore import (
-    QCoreApplication,
-    QDate,
-    QDateTime,
-    QLocale,
-    QMetaObject,
-    QObject,
-    QPoint,
-    QRect,
-    QSize,
-    Qt,
-    QTime,
-    QUrl,
-)
-from PySide6.QtGui import (
-    QBrush,
-    QColor,
-    QConicalGradient,
-    QCursor,
-    QFont,
-    QFontDatabase,
-    QGradient,
-    QIcon,
-    QImage,
-    QKeySequence,
-    QLinearGradient,
-    QPainter,
-    QPalette,
-    QPixmap,
-    QRadialGradient,
-    QTransform,
-)
-from PySide6.QtWidgets import (
-    QApplication,
-    QDialog,
-    QGridLayout,
-    QHBoxLayout,
-    QLabel,
-    QLayout,
-    QLineEdit,
-    QPushButton,
-    QSizePolicy,
-    QSpacerItem,
-    QVBoxLayout,
-    QWidget,
-)
-
+from PySide6.QtCore import (QCoreApplication, QDate, QDateTime, QLocale,
+    QMetaObject, QObject, QPoint, QRect,
+    QSize, QTime, QUrl, Qt)
+from PySide6.QtGui import (QBrush, QColor, QConicalGradient, QCursor,
+    QFont, QFontDatabase, QGradient, QIcon,
+    QImage, QKeySequence, QLinearGradient, QPainter,
+    QPalette, QPixmap, QRadialGradient, QTransform)
+from PySide6.QtWidgets import (QApplication, QDialog, QGridLayout, QHBoxLayout,
+    QLabel, QLayout, QLineEdit, QPushButton,
+    QSizePolicy, QSpacerItem, QVBoxLayout, QWidget)
 
 class Ui_Dialog(object):
     def setupUi(self, Dialog):
@@ -89,15 +52,16 @@ class Ui_Dialog(object):
 
         self.gridLayout.addWidget(self.wait_line_edit, 1, 1, 1, 1)
 
+
         self.verticalLayout.addLayout(self.gridLayout)
 
-        self.verticalSpacer = QSpacerItem(20, 40, QSizePolicy.Minimum, QSizePolicy.Expanding)
+        self.verticalSpacer = QSpacerItem(20, 40, QSizePolicy.Policy.Minimum, QSizePolicy.Policy.Expanding)
 
         self.verticalLayout.addItem(self.verticalSpacer)
 
         self.horizontalLayout_2 = QHBoxLayout()
         self.horizontalLayout_2.setObjectName(u"horizontalLayout_2")
-        self.horizontalSpacer_2 = QSpacerItem(40, 20, QSizePolicy.Expanding, QSizePolicy.Minimum)
+        self.horizontalSpacer_2 = QSpacerItem(40, 20, QSizePolicy.Policy.Expanding, QSizePolicy.Policy.Minimum)
 
         self.horizontalLayout_2.addItem(self.horizontalSpacer_2)
 
@@ -111,11 +75,12 @@ class Ui_Dialog(object):
 
         self.horizontalLayout_2.addWidget(self.change_state_button)
 
+
         self.verticalLayout.addLayout(self.horizontalLayout_2)
 
         self.horizontalLayout = QHBoxLayout()
         self.horizontalLayout.setObjectName(u"horizontalLayout")
-        self.horizontalSpacer = QSpacerItem(40, 20, QSizePolicy.Expanding, QSizePolicy.Minimum)
+        self.horizontalSpacer = QSpacerItem(40, 20, QSizePolicy.Policy.Expanding, QSizePolicy.Policy.Minimum)
 
         self.horizontalLayout.addItem(self.horizontalSpacer)
 
@@ -129,12 +94,13 @@ class Ui_Dialog(object):
 
         self.horizontalLayout.addWidget(self.cancel_button)
 
+
         self.verticalLayout.addLayout(self.horizontalLayout)
 
-        # if QT_CONFIG(shortcut)
+#if QT_CONFIG(shortcut)
         self.drives_label.setBuddy(self.drives_line_edit)
         self.wait_label.setBuddy(self.wait_line_edit)
-        # endif // QT_CONFIG(shortcut)
+#endif // QT_CONFIG(shortcut)
         QWidget.setTabOrder(self.drives_line_edit, self.wait_line_edit)
         QWidget.setTabOrder(self.wait_line_edit, self.log_button)
         QWidget.setTabOrder(self.log_button, self.change_state_button)
@@ -146,8 +112,8 @@ class Ui_Dialog(object):
         self.ok_button.setDefault(True)
         self.cancel_button.setDefault(False)
 
-        QMetaObject.connectSlotsByName(Dialog)
 
+        QMetaObject.connectSlotsByName(Dialog)
     # setupUi
 
     def retranslateUi(self, Dialog):
@@ -158,5 +124,5 @@ class Ui_Dialog(object):
         self.change_state_button.setText(QCoreApplication.translate("Dialog", u"Start/Stop", None))
         self.ok_button.setText(QCoreApplication.translate("Dialog", u"Ok", None))
         self.cancel_button.setText(QCoreApplication.translate("Dialog", u"Cancel", None))
-
     # retranslateUi
+
